- limit conversation messages to last 10 before passing to llm - prompt engineering - create a separate chat client console
