+++ src/adapter/llm.py
@@ -14,8 +14,9 @@ class LlmService:
 
         template = """
         You are an assistant for question-answering tasks. 
-        Use ONLY the following piece of context below to answer the question in detail and DO NOT add any information outside of 
-        it. Can you provide information about the \"document source\" and  \"page number\" if present at the end of your response as references.  
+        Use ONLY the following piece of context below to answer the question and DO NOT add any information outside of 
+        it. Provide concise overview or brief summary and mention the \"document source\" and  \"page number\" if present at the end of 
+        your response as references.  
         
         If context information provided is missing, just say you don't know, don't make up an answer.
